+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 0.2.2
 
+* `Repos`
+    * `Common`
+        * Several new methods `ReadOneToManyKeyValueRepo#getAll`
+        * Several new method `WriteOneToManyKeyValueRepo#add` and several extensions
+        * Several new method `WriteOneToManyKeyValueRepo#remove` and several extensions
+
 ## 0.2.1
 
 * `Pagination`